+++ tests/conftest.py
@@ -19,6 +19,9 @@ def simple_member_data() -> dict:
         "id": "6F63AF02CE05328153ABA477C76E6189",
         "lastName": "Gleason",
         "phoneNumber": "+123456789",
+        "profile": {
+            "id": "364C188137AD92DC0F32E1A31A0E1731",
+        },
     }
 
 
@@ -120,6 +123,9 @@ def complex_group_data() -> dict:
                 "id": "6F63AF02CE05328153ABA477C76E6189",
                 "lastName": "Gleason",
                 "phoneNumber": "+123456789",
+                "profile": {
+                    "id": "364C188137AD92DC0F32E1A31A0E1731",
+                },
                 "roles": [
                     "29A7724B47ABEE7B3C9DC347E13A50B4",
                 ],